refactor: move sidebar metadata support
